Rename PostgreSQL 'testdb' package to 'testdbpgx'
